Clean up  context.ts
