--- src/cca/app/polymer/browser-elements.html
@@ -1,3 +1,0 @@
-<!doctype html>
-<!-- Imports for all Chrome-specific Polymer elements. This ensures that the elements will be vulcanized along with the generic UI Polymer elements. -->
-
